+++ src/graphics/bmp.rs
@@ -5,31 +5,31 @@ use graphics::point::*;
 use graphics::size::*;
 
 pub struct BMP {
-    pub data: u32,
+    pub data: usize,
     pub size: Size
 }
 
 impl BMP {
-    pub unsafe fn new(file_data: u32) -> BMP {
+    pub unsafe fn new(file_data: usize) -> BMP {
         let data;
         let size;
         if file_data > 0
             && *(file_data as *const u8) == 'B' as u8
             && *((file_data + 1) as *const u8) == 'M' as u8
         {
-            let file_size = *((file_data + 0x2) as *const u32);
-            let offset = *((file_data + 0xA) as *const u32);
-            let width = *((file_data + 0x12) as *const u32);
-            let height = *((file_data + 0x16) as *const u32);
-            let depth = *((file_data + 0x1C) as *const u16) as u32;
+            let file_size = *((file_data + 0x2) as *const u32) as usize;
+            let offset = *((file_data + 0xA) as *const u32) as usize;
+            let width = *((file_data + 0x12) as *const u32) as usize;
+            let height = *((file_data + 0x16) as *const u32) as usize;
+            let depth = *((file_data + 0x1C) as *const u16) as usize;
             
             let bytes = (depth + 7)/8;
             let row_bytes = (depth * width + 31)/32 * 4;
             
             data = alloc(width * height * 4);
             size = Size {
-                width: width,
-                height: height
+                width: width as u32,
+                height: height as u32
             };
             for y in 0..height {
                 for x in 0..width {
@@ -75,7 +75,7 @@ impl BMP {
             && point.y >= 0
             && point.y < self.size.height as i32
         {
-            return *((self.data + (point.y as u32*self.size.width + point.x as u32)*4) as *const Color);
+            return *((self.data + (point.y as usize*self.size.width as usize + point.x as usize)*4) as *const Color);
         }else{
             return Color::new(0, 0, 0);
         }
